Renamed MarkeAPI to MarkeCore
