+++ MarketCore/MarketCore.csproj
